Add MapElement class
