+++ src/gm/maplabel.js
@@ -33,15 +33,19 @@
 
 goog.provide('olgm.gm.MapLabel');
 
+goog.require('olgm.gm.MapElement');
+
 
 /**
  * Creates a new Map Label
  * @constructor
- * @extends {google.maps.OverlayView}
+ * @extends {olgm.gm.MapElement}
  * @param {Object.<string, *>=} opt_options Optional properties to set.
  * @api
  */
 olgm.gm.MapLabel = function(opt_options) {
+  goog.base(this);
+
   this.set('font', 'normal 10px sans-serif');
   this.set('textAlign', 'center');
   this.set('textBaseline', 'middle');
@@ -50,30 +54,7 @@ olgm.gm.MapLabel = function(opt_options) {
 
   this.setValues(opt_options);
 };
-if (window.google && window.google.maps) {
-  goog.inherits(olgm.gm.MapLabel, google.maps.OverlayView);
-}
-
-
-/**
- * @type {boolean}
- * @private
- */
-olgm.gm.MapLabel.prototype.drawn_ = false;
-
-
-/**
- * @type {number}
- * @private
- */
-olgm.gm.MapLabel.prototype.height_ = 0;
-
-
-/**
- * @type {number}
- * @private
- */
-olgm.gm.MapLabel.prototype.width_ = 0;
+goog.inherits(olgm.gm.MapLabel, olgm.gm.MapElement);
 
 
 /**
@@ -165,107 +146,5 @@ olgm.gm.MapLabel.prototype.onAdd = function() {
   var panes = this.getPanes();
   if (panes) {
     panes.markerLayer.appendChild(canvas);
-  }
-};
-
-
-/**
- * Note: mark as `@api` to make the minimized version include this method.
- * @api
- */
-olgm.gm.MapLabel.prototype.draw = function() {
-
-  if (this.drawn_) {
-    this.redraw_();
-    return;
-  }
-
-  var canvas = this.canvas_;
-  if (!canvas) {
-    // onAdd has not been called yet.
-    return;
-  }
-
-  var ctx = canvas.getContext('2d');
-  var height = ctx.canvas.height;
-  var width = ctx.canvas.width;
-  this.width_ = width;
-  this.height_ = height;
-
-  if (!this.redraw_()) {
-    return;
-  }
-
-  this.drawn_ = true;
-};
-
-
-/**
- * Note: mark as `@api` to make the minimized version include this method.
- * @return {boolean} whether or not the function ran successfully
- * @private
- */
-olgm.gm.MapLabel.prototype.redraw_ = function() {
-  var latLng = /** @type {google.maps.LatLng} */ (this.get('position'));
-  if (!latLng) {
-    return false;
-  }
-
-  var projection = this.getProjection();
-  if (!projection) {
-    // The map projection is not ready yet so do nothing
-    return false;
-  }
-
-  var pos = projection.fromLatLngToDivPixel(latLng);
-  var height = this.height_;
-  var width = this.width_;
-
-  var offsetX = this.get('offsetX') || 0;
-  var offsetY = this.get('offsetY') || 0;
-  var style = this.canvas_.style;
-  style['top'] = (pos.y - (height / 2) + offsetY) + 'px';
-  style['left'] = (pos.x - (width / 2) + offsetX) + 'px';
-
-  style['visibility'] = this.getVisible_();
-
-  return true;
-};
-
-
-/**
- * Get the visibility of the label.
- * @private
- * @return {string} blank string if visible, 'hidden' if invisible.
- */
-olgm.gm.MapLabel.prototype.getVisible_ = function() {
-  var minZoom = /** @type {number} */ (this.get('minZoom'));
-  var maxZoom = /** @type {number} */ (this.get('maxZoom'));
-
-  if (minZoom === undefined && maxZoom === undefined) {
-    return '';
-  }
-
-  var map = this.getMap();
-  if (!map) {
-    return '';
-  }
-
-  var mapZoom = map.getZoom();
-  if (mapZoom < minZoom || mapZoom > maxZoom) {
-    return 'hidden';
-  }
-  return '';
-};
-
-
-/**
- * Note: mark as `@api` to make the minimized version include this method.
- * @api
- */
-olgm.gm.MapLabel.prototype.onRemove = function() {
-  var canvas = this.canvas_;
-  if (canvas && canvas.parentNode) {
-    canvas.parentNode.removeChild(canvas);
   }
 };
